+++ battle_tanks/client/headers/renderer/render_object.hpp
